chore: renamed comps
