+++ client/components/AllItems.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import SingleItem from './single-item'
+import SingleItem from './SingleItem'
 
 const AllItems = ({item}) => {
   return (
