+++ client/components/index.js
@@ -5,8 +5,8 @@
  */
 export {default as Main} from './main'
 export {default as UserHome} from './user-home'
-export {default as SingleItem} from './single-item'
-export {default as AllItems} from './all-items'
+export {default as SingleItem} from './SingleItem'
+export {default as AllItems} from './AllItems'
 export {Login, Signup} from './auth-form'
 export {default as ProductList} from './productlist'
 export {default as SingleUser} from './SingleUser'
